Update README.md
This assignment requires us to generate a random secure password application for employees, based on the criteria that they've selected.  The application runs in the browser and features updated HTML and CSS-powered JavaScript code. It uses a responsive user interface that adapts to multiple screen sizes. The password includes special characters, lowercase, uppercase, and numbers.